Change Opcode.getOpcodeByValue to accept an int

+++ dexlib2/src/main/java/org/jf/dexlib2/Opcode.java
@@ -328,7 +328,7 @@ public enum Opcode
         return opcodesByName.get(opcodeName.toLowerCase().hashCode());
     }
 
-    public static Opcode getOpcodeByValue(short opcodeValue) {
+    public static Opcode getOpcodeByValue(int opcodeValue) {
         return opcodesByValue[opcodeValue];
     }
